+++ src/main/java/com/tunisij/project2/handlers/YardHandler.java
@@ -8,7 +8,7 @@ public class YardHandler extends ConverterHandler {
 	
 	@Override
 	public String convert(String kilometers) {
-		return "Yards".equals(units) ? Double.toString(Integer.parseInt(kilometers) * 1093.61) : handler.convert(kilometers);
+		return "Yards".equals(units) ? Double.toString(Double.parseDouble(kilometers) * 1093.61) : handler.convert(kilometers);
 	}
 
 }
